build fixes
Moved classes & functions around,
Renamed Editors -> Editor
Builds successfully now
