+++ Assets/Editor/OTIEditorBase.cs
@@ -80,38 +80,6 @@ namespace oti.Editors
         public static string _WorldMonitoringExplanation = "Use this utility to give the agent awareness of events going on in the environment.";
 
         public static string _WorldMonitorSingleton = "The World Monitor: choose the parameters for your tracking system here.";
-
-        public static string[] _Alphabetic = new string[26] { "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z" };
-
-        /// <summary>
-        /// Assembles a continuous alphabetic unique string from index i: A (i = 0), ...., AB (i = 27), ....., BC (i = 54), ... AAB etc
-        /// </summary>
-        public static string _AlphabetAssembler(int i)
-        {
-            string s = "";
-            int wh = -1;
-
-            for (int j = 0; j <= i; j += _Alphabetic.Length)
-            {
-                wh += 1;
-            }
-
-            for (int j = 0; j <= wh; j++)
-            {
-                if (j > 0)
-                {
-                    j = wh + 1;
-                    s += _Alphabetic[i % _Alphabetic.Length];
-                }
-                else
-                {
-                    s += wh >= 1 ? _Alphabetic[wh - 1] : _Alphabetic[i % _Alphabetic.Length];
-                }
-            }
-
-            return s;
-        }
-
     }
 
     public struct OTIEditorFormat
