--- Assets/Editors.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ab9457df3437db54596ed4672f3b7ef6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
