+++ Assets/Materials/Black.mat.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f41e1ce644fdd374db4ebb859e1c966d
+guid: d6b586e5d1add28469f11f754f0df37f
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 2100000
