+++ Packages/manifest.json
@@ -1,0 +1,4 @@
+{
+	"dependencies": {
+	}
+}
